+++ ext/DynamicPPLMooncakeExt.jl
@@ -5,5 +5,8 @@ using Mooncake: Mooncake
 
 # This is purely an optimisation.
 Mooncake.@zero_derivative Mooncake.DefaultCtx Tuple{typeof(is_transformed),Vararg}
+Mooncake.@zero_derivative Mooncake.DefaultCtx Tuple{
+    typeof(DynamicPPL._get_range_and_linked),Vararg
+}
 
 end # module